BE email authentication

+++ web03-backend/src/main/java/com/web03backend/repositories/spec/ITestRepository.java
@@ -3,5 +3,5 @@ package com.web03backend.repositories.spec;
 import com.web03backend.domain.TestEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ITestRepository extends JpaRepository<TestEntity, Integer> {
+public interface ITestRepository extends JpaRepository<TestEntity, Long> {
 }
